+++ comics/comics/supereffective.py
@@ -12,6 +12,9 @@ class Crawler(CrawlerBase):
     history_capable_date = '2008-04-23'
     time_zone = -5
 
+    # Without User-Agent set, the server returns empty responses
+    headers = {'User-Agent': 'Mozilla/4.0'}
+
     def crawl(self, pub_date):
         url = 'http://www.vgcats.com/super/images/%s.gif' % (
             pub_date.strftime('%y%m%d'),)