+++ comics/comics/vgcats.py
@@ -14,6 +14,9 @@ class Crawler(CrawlerBase):
     history_capable_date = '2001-09-09'
     time_zone = -5
 
+    # Without User-Agent set, the server returns empty responses
+    headers = {'User-Agent': 'Mozilla/4.0'}
+
     def crawl(self, pub_date):
         # FIXME: Seems like they are using gif images now and then
         if pub_date < dt.date(2003, 5, 1):