+++ docs/changes.rst
@@ -46,6 +46,8 @@ Unreleased
 - Update ``eon`` with new source site.
 - Update ``lunch`` with new source site.
 - Update ``sometingpositive`` to get all releases.
+- Update ``supereffective`` to work after source site changes.
+- Update ``vgcats`` to work after source site changes.
 - Update ``yafgc`` to work after source site changes.